Introduce ConstellationBindMount
The former ConstellationMount was renamed ConstellationVolumeMount to reflect its inherent
assumption that the mount it wraps is a volume mount.

ConstellationBindMount is introduced, which allows constellation users to create bind
mounts that can be passed into ConstellationContainer without requiring irrelevant
volume details to be passed to the ConstellationBindMount.

The interfaces of the two ...Mount classes are made identical so that 'mounts', the
constructor parameter of ConstellationContainer, can be iterated over cleanly.

ConstellationVolumeMount now dictates the mount type as 'volume', to reflect its name;

previously, 'volume' could have been overridden with another type via kwargs (although
this would have caused an error when to_mount was called).